file_transfer_api/pyproject.toml
[tool.poetry]
name = "filetransferapi"
version = "0.1.0"
description = "A FastAPI based API"
authors = ["Ugo Okoroafor <39538470+Uokoroafor@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.105.0"
uvicorn = "^0.25.0"
python-multipart = "^0.0.6"
sqlalchemy = "^2.0.23"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
requests-mock = "^1.11.0"
coverage = "^7.4.1"
httpx = "^0.26.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"

[tool.poetry.scripts]
launch-api = "scripts.launch_api:main"
setup-local-db = "scripts.setup_local_db:main"


[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
